@@ -6,6 +6,7 @@ import '../../data/app_scm_engenharia_mobile_bll.dart';
 import '../../data/tb_environment_variable.dart';
 import '../../help/components.dart';
 import '../../help/formatter/cnpj_input_formatter.dart';
+import '../../help/parameter_result_view.dart';
 import '../../models/input/input_sici_fust_form_model.dart';
 import '../../models/operation.dart';
 import '../../models/output/output_environment_variables_model.dart';
@@ -24,17 +25,17 @@ class DadosEmServicosView extends StatefulWidget {
   DadosEmServicosState createState() => DadosEmServicosState();
 }
 
-class DadosEmServicosState extends State<DadosEmServicosView> {
+class DadosEmServicosState extends State<DadosEmServicosView> with ParameterResultViewEvent {
 
   TypeView statusView = TypeView.viewLoading;
   OutputEnvironmentVariablesModel resulEnvironmentVariables = OutputEnvironmentVariablesModel();
 
   //CNPJ:
-  final txtControllerCnpj =  TextEditingController();
-  final  focusNodeCnpj = FocusNode();
+  //final txtControllerCnpj =  TextEditingController();
+ // final  focusNodeCnpj = FocusNode();
 
-  List<UtilDropdownList> listMonths = <UtilDropdownList>[];
-  late UtilDropdownList utilDropdownListMonth;
+  //List<UtilDropdownList> listMonths = <UtilDropdownList>[];
+  //late UtilDropdownList utilDropdownListMonth;
   List<Uf>? ufDropdownList;
   Uf ufValue = Uf();
   List<TipoCliente>? customerTypeDropdownList;
@@ -48,7 +49,7 @@ class DadosEmServicosState extends State<DadosEmServicosView> {
   List<TipoProduto>? productTypeDropdownList;
   TipoProduto productTypeValue = TipoProduto();
   CodIbge valueCodIbge = CodIbge();
-  final txtNumberYear = TextEditingController();
+ // final txtNumberYear = TextEditingController();
   final txtCounty = TextEditingController();
    String hintTextCounty = 'O código IBGE..';
 
@@ -91,7 +92,7 @@ class DadosEmServicosState extends State<DadosEmServicosView> {
       sInput.quantidadeAcesso = txtControllerAccesses.text;
       Navigator.pop(context, sInput);
     } catch (error) {
-      OnAlertError(error.toString());
+      OnAlert.onAlertError(context, error.toString());
     }
   }
 
@@ -182,10 +183,10 @@ class DadosEmServicosState extends State<DadosEmServicosView> {
   void initState() {
     super.initState();
     Future.delayed(Duration.zero, () async {
-      listMonths = <UtilDropdownList>[];
-      listMonths = await Components.onMonths();
+     // listMonths = <UtilDropdownList>[];
+      //listMonths = await Components.onMonths();
       setState(() {
-        utilDropdownListMonth = listMonths.first;
+        //utilDropdownListMonth = listMonths.first;
       });
     });
     onInc();
@@ -223,7 +224,7 @@ class DadosEmServicosState extends State<DadosEmServicosView> {
       case TypeView.viewLoading:
         return GlobalView.viewPerformingSearch(maxHeight,context);
       case TypeView.viewErrorInformation:
-        return GlobalView.viewErrorInformation(maxHeight,GlobalScaffold.erroInformacao,context);
+        return GlobalView.viewErrorInformation(maxHeight,erroInformation,context);
       case TypeView.viewRenderInformation:
         return GlobalView.viewRenderSingleChildScrollView(maxHeight,Container( padding: const EdgeInsets.fromLTRB(15.0, 30.0, 15.0, 0.0),
             constraints: const BoxConstraints(
@@ -233,31 +234,7 @@ class DadosEmServicosState extends State<DadosEmServicosView> {
             mainAxisAlignment: MainAxisAlignment.start,
             crossAxisAlignment: CrossAxisAlignment.center,
             children: <Widget>[
-              Padding(padding: const EdgeInsets.fromLTRB(0.0, 20.0, 0.0, 10.0),child: TextField(
-                keyboardType: TextInputType.number,
-                controller: txtControllerCnpj,
-                focusNode: focusNodeCnpj,
-                textInputAction: TextInputAction.next,
-                style: const TextStyle(
-                    fontSize: 16,
-                    fontFamily: 'Poppins-Regular',
-                    fontWeight: FontWeight.w100,
-                    color: Color(0xFF323232)),
-                onSubmitted: (term) {
-                  focusNodeCnpj.unfocus();
-                  // FocusScope.of(context).requestFocus(focusNodeTelefoneMovel);
-                },
-                inputFormatters: [
-                  // obrigatório
-                  FilteringTextInputFormatter.digitsOnly,
-                  CnpjInputFormatter(),
-                ],
-                autofocus: false,
-                decoration: const InputDecoration(
-                  labelText: 'CNPJ:',
-                  hintText: '',
-                ),
-              ),),
+
               Padding(padding: const EdgeInsets.fromLTRB(0.0, 15.0, 0.0, 15.0),child:   DropdownButtonFormField<Uf>(
                 elevation: 7,
                 dropdownColor: AppThema.themeNotifierState.value.mode == ThemeMode.dark ? const Color(0xff000000) : const Color(0xffFFFFFF),
@@ -302,81 +279,7 @@ class DadosEmServicosState extends State<DadosEmServicosView> {
                   txtCounty.text = "";
                 },
               ),),
-              Padding(padding: const EdgeInsets.fromLTRB(0.0, 15.0, 0.0, 15.0),child:  Row(
-                children: [
-                  Expanded(
-                    child: TextField(
-                      autofocus: false,
-                      keyboardType: TextInputType.number,
-                      controller: txtNumberYear,
-                      textInputAction: TextInputAction.go,
-                      style: const TextStyle(
-                          fontSize: 16,
-                          fontFamily: 'Poppins-Regular',
-                          fontWeight: FontWeight.w100,
-                          color: Color(0xFF323232)),
-                      inputFormatters: [
-                        // obrigatório
-                        FilteringTextInputFormatter.digitsOnly,
-                        LengthLimitingTextInputFormatter(4),
-                      ],
-                      onSubmitted: (value) {
-                        FocusScope.of(context).requestFocus(FocusNode());
-                      },
-                      decoration: const InputDecoration(
-                        hintText: 'Digite Ano',
-                        labelText: 'Ano',
-                      ),
-                    ),
-                  ),
-                  const SizedBox(
-                    width: 15.0,
-                  ),
-                  Expanded(
-                    child:  DropdownButtonFormField<UtilDropdownList>(
-                      elevation: 7,
-                      dropdownColor: AppThema.themeNotifierState.value.mode == ThemeMode.dark ? const Color(0xff000000) : const Color(0xffFFFFFF),
-                      isExpanded: true,
-                      isDense: true,
-                      icon: const Icon(
-                        Icons.expand_more,
-                        size: 23,
-                        color: Color(0xFFb8b8b8),
-                      ),
-                      decoration:  const InputDecoration(
-                        filled: true,
-                        labelText: 'Mês',
-                        hintText: 'Mês',
-                        //contentPadding: const EdgeInsets.fromLTRB(10.0, 18.0, 18.0, 0.0),
-                        border: InputBorder.none,
-                        //focusColor: Colors.transparent,
-                      ),
-                      value: utilDropdownListMonth,
-                      items: listMonths.map(
-                            (v) => DropdownMenuItem<UtilDropdownList>(
-                            value: v,
-                            child: Text(
-                              v.txt!,
-                              overflow: TextOverflow.ellipsis,
-                              softWrap: true,
-                              maxLines: 1,
-                              style: const TextStyle(
-                      fontSize: 16,
-                      fontFamily: 'Poppins-Regular',
-                      fontWeight: FontWeight.w100,
-                      color: Color(0xFF323232)),
 
-                            )),
-                      ).toList(),
-                      onChanged: (newValue) {
-                        setState(() {
-                          utilDropdownListMonth = newValue!;
-                        });
-                      },
-                    ),
-                  ),
-                ],
-              )),
               Padding( padding: const EdgeInsets.fromLTRB(0.0, 10.0, 0.0, 10.0),child:TextField(
                 onTap: () {
                   FocusScope.of(context).requestFocus(FocusNode());
@@ -418,7 +321,7 @@ class DadosEmServicosState extends State<DadosEmServicosView> {
                       });
                     }
                   } catch (error) {
-                    OnAlertError(error.toString());
+                    OnAlert.onAlertError(context, error.toString());
                   }},
                  onChanged: (value) {
                    FocusScope.of(context).requestFocus(FocusNode());
@@ -692,6 +595,7 @@ class DadosEmServicosState extends State<DadosEmServicosView> {
                       textAlign: TextAlign.start,
                       keyboardType: TextInputType.number,
                       focusNode: txtFocusNodeAccesses,
+                      textInputAction: TextInputAction.none,
                       style: const TextStyle(
                           fontSize: 16,
                           fontFamily: 'Poppins-Regular',
@@ -699,7 +603,6 @@ class DadosEmServicosState extends State<DadosEmServicosView> {
                           color: Color(0xFF323232)),
                       onSubmitted: (term) {
                         txtFocusNodeAccesses!.unfocus();
-                        onAdd();
                       },
                       decoration: const InputDecoration(
                         labelText: 'ACESSOS',
@@ -729,6 +632,8 @@ class DadosEmServicosState extends State<DadosEmServicosView> {
               ),),),
             ],
           ),),context);
+      case TypeView.viewThereIsNoInternet:
+        // TODO: Handle this case.
     }
   }
 }
