@@ -9,8 +9,8 @@ class CurrencyPtBrInputFormatter extends TextInputFormatter {
     }
 
     double value = double.parse(newValue.text);
-    final formatter = new NumberFormat("#,##0.00", "pt_BR");
-    String newText = "R\$ " + formatter.format(value/100);
+    final formatter = NumberFormat("#,##0.00", "pt_BR");
+    String newText = "R\$ ${formatter.format(value/100)}";
 
     return newValue.copyWith(
         text: newText,
