@@ -10,6 +10,7 @@ import 'dart:async';
 import 'dart:io';
 import 'package:http/http.dart' as http;
 
+import '../thema/app_thema.dart';
 import 'help_views/global_scaffold.dart';
 
 
@@ -73,11 +74,11 @@ class InicioState extends State<InicioView>  {
               text: TextSpan(children: [
                 TextSpan(
                     text: 'Licenças de TV e telefonia para  seu \r\n',
-                   style: Theme.of(context).textTheme.bodyText1?.copyWith(fontSize: 20,color:  Colors.white,),
+                   style: StylesThemas.textStyleTextTitle().copyWith(fontSize: 20,color:  Colors.white,),
                 ),
                 TextSpan(
                     text: 'Provedor. ',
-                    style: Theme.of(context).textTheme.bodyText1?.copyWith( fontSize: 25,color:  Colors.white, fontWeight: FontWeight.w800,),
+                    style: StylesThemas.textStyleTextTitle().copyWith( fontSize: 25,color:  Colors.white, fontWeight: FontWeight.w800,),
                 ),
               ])),),
         ),
@@ -182,7 +183,7 @@ class InicioState extends State<InicioView>  {
                 child: Column(
                   children:  <Widget>[
                     const Icon(Icons.search_outlined, color: Color(0xffef7d00),),
-                    Padding( padding: const EdgeInsets.only(top: 10.0),child: Text('Análise',  style: Theme.of(context).textTheme.bodyText1?.copyWith(fontSize: 12,color:  Colors.black87, fontWeight: FontWeight.w500,))),
+                    Padding( padding: const EdgeInsets.only(top: 10.0),child: Text('Análise',  style: StylesThemas.textStyleTextTitle().copyWith(fontSize: 12,color:  Colors.black87, fontWeight: FontWeight.w500,))),
                   ],
                 ),
               ),
@@ -193,7 +194,7 @@ class InicioState extends State<InicioView>  {
                 child: Column(
                   children: <Widget>[
                     const Icon(Icons.warning, color: Color(0xffef7d00),),
-                    Padding( padding: const EdgeInsets.only(top: 10.0),child: Text('Alertas',  style: Theme.of(context).textTheme.bodyText1?.copyWith(fontSize: 12,color:  Colors.black87, fontWeight: FontWeight.w500,))),
+                    Padding( padding: const EdgeInsets.only(top: 10.0),child: Text('Alertas',  style: StylesThemas.textStyleTextTitle().copyWith(fontSize: 12,color:  Colors.black87, fontWeight: FontWeight.w500,))),
 
                   ],
                 ),
@@ -205,7 +206,7 @@ class InicioState extends State<InicioView>  {
                 child: Column(
                   children: <Widget>[
                     Icon(Icons.description, color: Color(0xffef7d00),),
-                    Padding( padding: EdgeInsets.only(top: 10.0),child: Text('Recibos',   style: Theme.of(context).textTheme.bodyText1?.copyWith(fontSize: 12,color:  Colors.black87, fontWeight: FontWeight.w500,))),
+                    Padding( padding: EdgeInsets.only(top: 10.0),child: Text('Recibos',   style: StylesThemas.textStyleTextTitle().copyWith(fontSize: 12,color:  Colors.black87, fontWeight: FontWeight.w500,))),
 
                   ],
                 ),
@@ -217,7 +218,7 @@ class InicioState extends State<InicioView>  {
                 child: Column(
                   children:  <Widget>[
                     const Icon(Icons.inventory, size: 25, color: Color(0xffef7d00),),
-                    Padding( padding: const EdgeInsets.only(top: 10.0),child: Text('Documentos',  style: Theme.of(context).textTheme.bodyText1?.copyWith(fontSize: 12,color:  Colors.black87, fontWeight: FontWeight.w500,))),
+                    Padding( padding: const EdgeInsets.only(top: 10.0),child: Text('Documentos',  style: StylesThemas.textStyleTextTitle().copyWith(fontSize: 12,color:  Colors.black87, fontWeight: FontWeight.w500,))),
 
                   ],
                 ),
@@ -247,6 +248,7 @@ class InicioState extends State<InicioView>  {
                 style:  TextStyle(
                   fontWeight: FontWeight.w400,
                   fontSize: 20,
+                  color: Colors.white
                 ),
               ),
             ),
