@@ -2,6 +2,7 @@ import 'dart:async';
 import 'package:connectivity_plus/connectivity_plus.dart';
 import 'package:flutter/material.dart';
 import '../../help/navigation_service/route_paths.dart' as routes;
+import '../../help/parameter_result_view.dart';
 import '../../models/operation.dart';
 import '../../models/notification_models/notification_model.dart';
 import '../../thema/app_thema.dart';
@@ -18,7 +19,7 @@ class NotificacaoDetalhesView extends StatefulWidget {
   NotificationState createState() => NotificationState();
 }
 
-class NotificationState extends State<NotificacaoDetalhesView> {
+class NotificationState extends State<NotificacaoDetalhesView> with ParameterResultViewEvent {
 
   NotificationScmEngineering notificationScmEngineering = NotificationScmEngineering();
   late StreamSubscription<ConnectivityResult> subscription;
@@ -52,10 +53,10 @@ class NotificationState extends State<NotificacaoDetalhesView> {
       setState(() {
         if (notificationScmEngineering.titulo!.isEmpty) {
           statusView = TypeView.viewRenderInformation;
-          OnAlertError(error.toString());
+          OnAlert.onAlertError(context, error.toString());
         } else {
           statusView = TypeView.viewErrorInformation;
-          GlobalScaffold.erroInformacao = error.toString();
+          erroInformation = error.toString();
         }
       });
     }
@@ -132,7 +133,11 @@ class NotificationState extends State<NotificacaoDetalhesView> {
       case TypeView.viewLoading:
         return GlobalView.viewPerformingSearch(maxHeight,context);
       case TypeView.viewErrorInformation:
-        return GlobalView.viewErrorInformation(maxHeight,GlobalScaffold.erroInformacao,context);
+        return GlobalView.viewErrorInformation(maxHeight,erroInformation,context);
+      case TypeView.viewRenderInformation:
+        // TODO: Handle this case.
+      case TypeView.viewThereIsNoInternet:
+        // TODO: Handle this case.
     }
   }
 
