@@ -85,7 +85,7 @@ class AlterarSenhaState extends State<AlterarSenhaView> {
       }
     } catch (error) {
       OnRealizandoOperacao('',context);
-      OnAlertError(error.toString());
+      OnAlert.onAlertError(context,error.toString());
     }
   }
 
@@ -146,7 +146,7 @@ class AlterarSenhaState extends State<AlterarSenhaView> {
                 children: <Widget>[
                   Padding(padding: const EdgeInsets.fromLTRB(20.0, 25.0, 20.0, 10.0),child: Text(
                     'Alterar senha',
-                    style: Theme.of(context).textTheme.bodyText1?.copyWith(fontSize: 20,color:  Colors.black, fontWeight: FontWeight.w600,),
+                    style: StylesThemas.textStyleTextTitle().copyWith(fontSize: 20,color:  Colors.black, fontWeight: FontWeight.w600,),
                   ),),
                   const Padding(padding: EdgeInsets.fromLTRB(20.0, 10.0, 20.0, 20.0),child: Divider(color:Colors.black54),),
                   Padding(padding: const EdgeInsets.fromLTRB(0.0, 10.0, 0.0, 20.0),child:  TextField(
