@@ -85,7 +85,7 @@ class ConfiguracoesState extends State<Configuracoesview> {
             children: <Widget>[
               Padding(padding: const EdgeInsets.fromLTRB(20.0, 25.0, 20.0, 10.0),child: Text(
                 'Configurações',
-                style: Theme.of(context).textTheme.bodyText1?.copyWith(fontSize: 20,color:  Colors.black, fontWeight: FontWeight.w600,),
+                style: StylesThemas.textStyleTextTitle().copyWith(fontSize: 20,color:  Colors.black, fontWeight: FontWeight.w600,),
               ),),
               const Padding(padding: EdgeInsets.fromLTRB(20.0, 10.0, 20.0, 20.0),child: Divider(color:Colors.black54),),
               Padding(padding: const EdgeInsets.fromLTRB(0.0, 10.0, 0.0, 15.0),child:  ListTile(
