@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:url_launcher/url_launcher.dart';
 import '../../help/components.dart';
 import '../../help/navigation_service/route_paths.dart' as routes;
+import '../../help/parameter_result_view.dart';
 import '../../models/info_app.dart';
 import '../../thema/app_thema.dart';
 import '../help_views/global_scaffold.dart';
@@ -14,7 +15,7 @@ class SobreView extends StatefulWidget {
   SobreState createState() => SobreState();
 }
 
-class SobreState extends State<SobreView> {
+class SobreState extends State<SobreView> with ParameterResultViewEvent {
 
   TypeView statusView = TypeView.viewLoading;
   InfoApp infoApp = InfoApp();
@@ -90,7 +91,7 @@ class SobreState extends State<SobreView> {
       case TypeView.viewLoading:
         return GlobalView.viewPerformingSearch(maxHeight,context);
       case TypeView.viewErrorInformation:
-        return GlobalView.viewErrorInformation(maxHeight,GlobalScaffold.erroInformacao,context);
+        return GlobalView.viewErrorInformation(maxHeight,erroInformation,context);
       case TypeView.viewRenderInformation:
         return SingleChildScrollView(
             padding: const EdgeInsets.only(top: 10.0, right: 10.0, left: 10.0, bottom: 10.0),
@@ -115,7 +116,7 @@ class SobreState extends State<SobreView> {
                     children: <Widget>[
                       Padding(padding: const EdgeInsets.fromLTRB(20.0, 25.0, 20.0, 10.0),child: Text(
                         'Sobre',
-                        style: Theme.of(context).textTheme.bodyText1?.copyWith(fontSize: 20,color:  Colors.black, fontWeight: FontWeight.w600,),
+                        style: StylesThemas.textStyleTextTitle().copyWith(fontSize: 20,color:  Colors.black, fontWeight: FontWeight.w600,),
                       ),),
                       const Padding(padding: EdgeInsets.fromLTRB(20.0, 10.0, 20.0, 20.0),child: Divider(color:Colors.black54),),
                       Column(
@@ -131,11 +132,11 @@ class SobreState extends State<SobreView> {
                               text: TextSpan(children: [
                                 TextSpan(
                                     text: 'Nome Aplicativo :',
-                                    style: Theme.of(context).textTheme.bodyText1?.copyWith(fontSize: 18)
+                                    style: StylesThemas.textStyleTextTitle().copyWith(fontSize: 18)
                                 ),
                                 TextSpan(
                                     text: '  ${infoApp.appName!}  ',
-                                    style: Theme.of(context).textTheme.bodyText2?.copyWith(fontSize: 20,color: Color(0xff262626))
+                                    style: StylesThemas.textStyleTextTitle().copyWith(fontSize: 20,color: Color(0xff262626))
                                 ),
                               ])),),
                           Padding( padding: const EdgeInsets.only(top: 10.0, bottom: 10.0) ,child: RichText(
@@ -146,11 +147,11 @@ class SobreState extends State<SobreView> {
                               text: TextSpan(children: [
                                 TextSpan(
                                     text: 'Build : ',
-                                    style: Theme.of(context).textTheme.bodyText1?.copyWith(fontSize: 18)
+                                    style: StylesThemas.textStyleTextTitle().copyWith(fontSize: 18)
                                 ),
                                 TextSpan(
                                     text: '  ${infoApp.buildNumber!}  ',
-                                    style: Theme.of(context).textTheme.bodyText2?.copyWith(fontSize: 20,color: Color(0xff262626))
+                                    style: StylesThemas.textStyleTextTitle().copyWith(fontSize: 20,color: Color(0xff262626))
                                 ),
                               ])),),
                           Padding( padding: const EdgeInsets.only(top: 10.0, bottom: 10.0) ,child: InkWell(
@@ -162,11 +163,11 @@ class SobreState extends State<SobreView> {
                                 text: TextSpan(children: [
                                   TextSpan(
                                       text: 'Versão :  ',
-                                      style: Theme.of(context).textTheme.bodyText1?.copyWith(fontSize: 18)
+                                      style: StylesThemas.textStyleTextTitle().copyWith(fontSize: 18)
                                   ),
                                   TextSpan(
                                       text: '  ${infoApp.version!}  ',
-                                      style: Theme.of(context).textTheme.bodyText2?.copyWith(fontSize: 20,color: Color(0xff262626))
+                                      style: StylesThemas.textStyleTextTitle().copyWith(fontSize: 20,color: Color(0xff262626))
                                   ),
                                 ])),
                             onTap: () async {
@@ -181,11 +182,11 @@ class SobreState extends State<SobreView> {
                               text: TextSpan(children: [
                                 TextSpan(
                                     text: 'Dispositivo : ',
-                                    style: Theme.of(context).textTheme.bodyText1?.copyWith(fontSize: 18)
+                                    style: StylesThemas.textStyleTextTitle().copyWith(fontSize: 18)
                                 ),
                                 TextSpan(
                                     text: '  ${infoApp.deviceVersion!}  ',
-                                    style: Theme.of(context).textTheme.bodyText2?.copyWith(fontSize: 20,color: Color(0xff262626))
+                                    style: StylesThemas.textStyleTextTitle().copyWith(fontSize: 20,color: Color(0xff262626))
                                 ),
                               ])),),
                           Padding( padding: const EdgeInsets.only(top: 10.0, bottom: 10.0) ,child:  RichText(
@@ -196,11 +197,11 @@ class SobreState extends State<SobreView> {
                               text: TextSpan(children: [
                                 TextSpan(
                                     text: 'ID do Dispositivo : ',
-                                    style: Theme.of(context).textTheme.bodyText1?.copyWith(fontSize: 18)
+                                    style: StylesThemas.textStyleTextTitle().copyWith(fontSize: 18)
                                 ),
                                 TextSpan(
                                     text: '  ${infoApp.idDevice!}  ',
-                                    style: Theme.of(context).textTheme.bodyText2?.copyWith(fontSize: 20,color: Color(0xff262626))
+                                    style: StylesThemas.textStyleTextTitle().copyWith(fontSize: 20,color: Color(0xff262626))
                                 ),
                               ])),),
                           Center(child: Container(
@@ -221,11 +222,11 @@ class SobreState extends State<SobreView> {
                                           TextSpan(
                                             text:
                                             'Este canal é exclusivo para dúvidas, sugestões e reclamações sobre o ',
-                                            style: Theme.of(context).textTheme.bodyText1?.copyWith(fontSize: 18),
+                                            style: StylesThemas.textStyleTextTitle().copyWith(fontSize: 18),
                                           ),
                                           TextSpan(
                                             text: 'APLICATIVO',
-                                            style: Theme.of(context).textTheme.bodyText2?.copyWith(fontSize: 18),
+                                            style: StylesThemas.textStyleTextTitle().copyWith(fontSize: 18),
                                           ),
                                         ])),
                                   ),
@@ -265,6 +266,8 @@ class SobreState extends State<SobreView> {
                       ),
                     ],
                   ),)),));
+      case TypeView.viewThereIsNoInternet:
+        // TODO: Handle this case.
     }
   }
 }
