@@ -2,6 +2,7 @@ import 'package:connectivity_plus/connectivity_plus.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import '../../help/formatter/cpf_input_formatter.dart';
+import '../../help/parameter_result_view.dart';
 import '../../help/responsive.dart';
 import '../../models/operation.dart';
 import '../../thema/app_thema.dart';
@@ -18,7 +19,7 @@ class CancelarAcessoView extends StatefulWidget {
   CancelarAcessoState createState() => CancelarAcessoState();
 }
 
-class CancelarAcessoState extends State<CancelarAcessoView> {
+class CancelarAcessoState extends State<CancelarAcessoView> with ParameterResultViewEvent {
 
 
   TextEditingController txtControllerUsuario = TextEditingController();
@@ -26,7 +27,8 @@ class CancelarAcessoState extends State<CancelarAcessoView> {
   onCancelarAcesso() async {
     try {
       if (await Connectivity().checkConnectivity() == ConnectivityResult.none) {
-        OnAlertError('Parece que você está sem internet ! Por favor, verifique a sua conexão e tente novamente.');
+        OnAlert.onAlertError(context, "Parece que você está sem internet ! Por favor, verifique a sua conexão e tente novamente.");
+
       } else {
         if (txtControllerUsuario.text.isEmpty) {
           throw ('Credencial e obrigatório\n');
@@ -45,13 +47,13 @@ class CancelarAcessoState extends State<CancelarAcessoView> {
         }
       }
     } catch (error) {
-      OnAlertError(error.toString().split('\r\n').first);
+      OnAlert.onAlertError(context, error.toString().split('\r\n').first);
     }
   }
 
   onInc() async {
     try {
-      GlobalScaffold.erroInformacao = '';
+      erroInformation = '';
       if (await Connectivity().checkConnectivity() == ConnectivityResult.none)
       {
         GlobalScaffold.instance.navigatorKey.currentState?.pushNamed(
@@ -133,7 +135,7 @@ class CancelarAcessoState extends State<CancelarAcessoView> {
                       TextSpan(
                         text:
                         'Prezado cliente,\r\n',
-                        style: Theme.of(context).textTheme.bodyText1?.copyWith(fontSize: 20,color:  Colors.black, fontWeight: FontWeight.w600,),
+                        style: StylesThemas.textStyleTextTitle().copyWith(fontSize: 20,color:  Colors.black, fontWeight: FontWeight.w600,),
                       ),
                       const TextSpan(
                         text: 'O cancelamento do serviço de participação implica no não-recebimento de relatórios pela Internet e revoga seu acesso. Para confirmar o cancelamento, informe sua credencial  e clique no botão "CONFIRMAR"',
