@@ -10,6 +10,7 @@ import '../../data/tb_user.dart';
 import '../../help/navigation_service/route_paths.dart' as routes;
 import '../../models/operation.dart';
 import '../../models/user_response_model.dart';
+import '../../thema/app_thema.dart';
 import '../../web_service/servico_mobile_service.dart';
 import '../help_views/global_scaffold.dart';
 import 'package:scm_engenharia_app/models/global_user_logged.dart' as global_user_logged;
@@ -88,7 +89,7 @@ class LoginState extends State<LoginView> {
       }
     } catch (error) {
       OnRealizandoOperacao('',context);
-      OnAlertError(error.toString());
+      OnAlert.onAlertError(context,error.toString());
     }
   }
 
@@ -98,6 +99,9 @@ class LoginState extends State<LoginView> {
     // NotificationHandler().subscribeToTopic("scmengenhariaUserNLogado");
     //txtControllerEmail.text = "rodolforomao@gmail.com";
     //txtControllerPassword.text = "1234567";
+
+     //txtControllerEmail.text = "fernando.oliveira@scmengenharia.com.br";
+     //txtControllerPassword.text = "trabalho@4030";
   }
 
   @override
@@ -151,11 +155,11 @@ class LoginState extends State<LoginView> {
                   text: TextSpan(children: [
                     TextSpan(
                       text: 'Olá, \r\n',
-                      style: Theme.of(context).textTheme.bodyText1?.copyWith(fontSize: 22,color:  Colors.black54,),
+                      style: StylesThemas.textStyleTextTitle().copyWith(fontSize: 22,color:  Colors.black54,),
                     ),
                     TextSpan(
                       text: 'digite seu e-mail e senha',
-                      style: Theme.of(context).textTheme.bodyText1?.copyWith(fontSize: 22,color:  Colors.black54,),
+                      style: StylesThemas.textStyleTextTitle().copyWith(fontSize: 22,color:  Colors.black54,),
                     ),
                   ])),),
               Padding(padding: const EdgeInsets.fromLTRB(0.0,10.0,0.0,10.0),child:TextField(
