@@ -1,40 +1,39 @@
 import 'dart:async';
 import 'dart:convert';
 import 'dart:io';
-import 'package:flutter/foundation.dart';
+import 'dart:typed_data';
 import 'package:http/http.dart' as http;
-import 'package:path_provider/path_provider.dart';
 import '../data/JWTTokenDbLocal.dart';
 import '../help/components.dart';
 import '../models/operation.dart';
 import '../models/input/input_sici_fust_form_model.dart';
+import 'package:http_parser/http_parser.dart';
 
 class ServicoMobileService {
-  static var Url = "http://dici.scmengenharia.com.br";
+  static var Url = "https://dici.scmengenharia.com.br";
   //static const Url = "http://10.0.2.2:8083";
-  static final UrlDebug = "http://10.200.4.77:8083";
+  //static final Url = "http://10.200.4.77:8083";
   //static final Url = "http://wsscm.ddns.net";
 
   static Future<Operation> onLogin(String usuario, String password) async {
     Operation operacao = Operation();
     try {
-      String? token = await Components.JWTToken(usuario, password);
-
+      String? token = Components.JWTToken(usuario, password);
       final response = await http
-          .post(Uri.parse((kDebugMode ? Url : UrlDebug) + "/login_ws"),
-              headers: {
-                //"Content-type": "multipart/form-data",
-                'token': token!,
-              },
-              encoding: Encoding.getByName('utf-8'))
-          .timeout(const Duration(seconds: 50));
+          .post(Uri.parse("$Url/login_ws"),
+          headers: {
+            //"Content-type": "multipart/form-data",
+            'token': token!,
+          },
+          encoding: Encoding.getByName('utf-8'));
+          //.timeout(const Duration(seconds: 50));
       operacao.statusCode = response.statusCode;
       if (response.statusCode == 200) {
         if (!response.body.isNotEmpty) {
           throw (ApiRestInformation.problemOfComunication);
         } else {
           Map<String, dynamic> map =
-              jsonDecode(Components.removeAllHtmlTags(response.body));
+          jsonDecode(Components.removeAllHtmlTags(response.body));
           OperationJson resp = OperationJson.fromJson(map);
           operacao.erro = !resp.status!;
           operacao.message = resp.message;
@@ -70,16 +69,10 @@ class ServicoMobileService {
     Operation operacao = Operation();
     try {
       String? token = await ComponentsJWTToken.JWTTokenPadrao();
-      Map<String, String> headers = {
-        'Content-Type': 'application/json; charset=utf-8',
-        'token': token!,
-      };
       http.MultipartRequest response;
       response = http.MultipartRequest(
-          'POST',
-          Uri.parse(
-              (kDebugMode ? Url : UrlDebug) + "/usuario/alterar_senha_ws"));
-      response.headers.addAll(headers);
+          'POST', Uri.parse("$Url/usuario/alterar_senha_ws"));
+      response.headers.addAll(ApiRestInformation.onHeadersToken(token!));
       response.fields['nova_senha'] = senha;
       var streamedResponse = await response.send();
       final respStr = await streamedResponse.stream.bytesToString();
@@ -89,7 +82,7 @@ class ServicoMobileService {
           throw (ApiRestInformation.problemOfComunication);
         } else {
           Map<String, dynamic> map =
-              jsonDecode(Components.removeAllHtmlTags(respStr));
+          jsonDecode(Components.removeAllHtmlTags(respStr));
           OperationJson resp = OperationJson.fromJson(map);
           operacao.erro = !resp.status!;
           operacao.message = resp.message;
@@ -125,15 +118,8 @@ class ServicoMobileService {
     Operation operacao = Operation();
     try {
       String? token = await ComponentsJWTToken.JWTTokenPadrao();
-      Map<String, String> headers = {
-        'Content-Type': 'application/json; charset=utf-8',
-        'token': token!,
-      };
       http.MultipartRequest response;
-      response = http.MultipartRequest(
-          'POST',
-          Uri.parse(
-              (kDebugMode ? Url : UrlDebug) + "/usuario/alterar_senha_ws"));
+      response = http.MultipartRequest('POST', Uri.parse("$Url/usuario/alterar_senha_ws"));
       response.headers.addAll(ApiRestInformation.onHeadersToken(token!));
       response.fields['nova_senha'] = senha;
       var streamedResponse = await response.send();
@@ -144,7 +130,7 @@ class ServicoMobileService {
           throw (ApiRestInformation.problemOfComunication);
         } else {
           Map<String, dynamic> map =
-              jsonDecode(Components.removeAllHtmlTags(respStr));
+          jsonDecode(Components.removeAllHtmlTags(respStr));
           OperationJson resp = OperationJson.fromJson(map);
           operacao.erro = !resp.status!;
           operacao.message = resp.message;
@@ -180,15 +166,10 @@ class ServicoMobileService {
     Operation operacao = Operation();
     try {
       String? token = await ComponentsJWTToken.JWTTokenPadrao();
-      final response = await http
-          .post(
-              Uri.parse((kDebugMode ? Url : UrlDebug) +
-                  "/analise/Analise/recuperarVariaveisAmbiente_ws"),
-              body: null,
-              headers: ApiRestInformation.onHeadersToken(token!),
-              encoding: Encoding.getByName("utf-8"))
-          .timeout(const Duration(seconds: 10));
-
+      final response = await http.post(Uri.parse("$Url/analise/Analise/recuperarVariaveisAmbiente_ws"),
+          body: null,
+          headers: ApiRestInformation.onHeadersToken(token!),
+          encoding: Encoding.getByName("utf-8"));//.timeout(const Duration(seconds: 10));
       operacao.erro = false;
       operacao.message = "Operação realizada com sucesso";
       operacao.result = null;
@@ -196,8 +177,7 @@ class ServicoMobileService {
         if (!response.body.isNotEmpty) {
           throw (ApiRestInformation.problemOfComunication);
         } else {
-          Map<String, dynamic> map =
-              jsonDecode(Components.removeAllHtmlTags(response.body));
+          Map<String, dynamic> map = jsonDecode(Components.removeAllHtmlTags(response.body));
           OperationJson resp = OperationJson.fromJson(map);
           operacao.erro = !resp.status!;
           operacao.message = resp.message;
@@ -224,16 +204,10 @@ class ServicoMobileService {
     Operation operacao = Operation();
     try {
       String? token = await ComponentsJWTToken.JWTTokenPadrao();
-      Map<String, String> headers = {
-        'Content-Type': 'application/json; charset=utf-8',
-        'token': token!,
-      };
       http.MultipartRequest response;
       response = http.MultipartRequest(
-          'POST',
-          Uri.parse((kDebugMode ? Url : UrlDebug) +
-              "/administracao/bloquear_usuario_ws"));
-      response.headers.addAll(headers);
+          'POST', Uri.parse("$Url/administracao/bloquear_usuario_ws"));
+      response.headers.addAll(ApiRestInformation.onHeadersToken(token!));
       response.fields['id'] = id;
       response.fields['acao'] = 'true';
       var streamedResponse = await response.send();
@@ -244,7 +218,7 @@ class ServicoMobileService {
           throw (ApiRestInformation.problemOfComunication);
         } else {
           Map<String, dynamic> map =
-              jsonDecode(Components.removeAllHtmlTags(respStr));
+          jsonDecode(Components.removeAllHtmlTags(respStr));
           OperationJson resp = OperationJson.fromJson(map);
           operacao.erro = !resp.status!;
           operacao.message = resp.message;
@@ -276,61 +250,55 @@ class ServicoMobileService {
     return operacao;
   }
 
-  static Future<Operation> onMakeReleasesSici(
-      InputSiciFileModel siciFileModel) async {
+  static Future<Operation> onMakeReleasesSici(InputSiciFileModel siciFileModel) async {
     Operation operacao = Operation();
     try {
       operacao.erro = false;
       operacao.message = 'Operação realizada com sucesso';
       operacao.result = null;
       String? token = await ComponentsJWTToken.JWTTokenPadrao();
-      Map<String, String> headers = {
-        'Content-Type': 'application/json; charset=utf-8',
-        "token": token!,
-      };
       http.MultipartRequest response;
-      response = http.MultipartRequest('POST',
-          Uri.parse((kDebugMode ? Url : UrlDebug) + "/analise/lancamento_ws"));
-      response.headers.addAll(headers);
+      response = http.MultipartRequest('POST', Uri.parse("$Url/analise/lancamento_ws"));
+      response.headers.addAll(ApiRestInformation.onHeadersToken(token!));
       if (siciFileModel.id!.isNotEmpty) {
         response.fields['controllerId'] = (siciFileModel.id ?? "");
       }
       response.fields['controllerPeriodoReferencia'] =
-          (siciFileModel.periodoReferencia ?? "");
+      (siciFileModel.periodoReferencia ?? "");
       response.fields['controllerPeriodoReferencia'] =
-          (siciFileModel.periodoReferencia ?? "");
+      (siciFileModel.periodoReferencia ?? "");
       response.fields['controllerRazaoSocial'] =
-          (siciFileModel.razaoSocial ?? "");
+      (siciFileModel.razaoSocial ?? "");
       response.fields['controllerTelefoneFixo'] =
-          siciFileModel.telefoneFixo == null ? "" : siciFileModel.telefoneFixo!;
+      siciFileModel.telefoneFixo == null ? "" : siciFileModel.telefoneFixo!;
       response.fields['controllerCNPJ'] =
-          siciFileModel.cnpj == null ? "" : siciFileModel.cnpj!;
+      siciFileModel.cnpj == null ? "" : siciFileModel.cnpj!;
       response.fields['controllerTelefoneCelular'] =
-          siciFileModel.telefoneMovel == null
-              ? ""
-              : siciFileModel.telefoneMovel!;
+      siciFileModel.telefoneMovel == null
+          ? ""
+          : siciFileModel.telefoneMovel!;
       response.fields['controllerReceitaBruta'] =
-          siciFileModel.receitaBruta == null ? "" : siciFileModel.receitaBruta!;
+      siciFileModel.receitaBruta == null ? "" : siciFileModel.receitaBruta!;
       response.fields['controllerAliqSimples'] =
-          siciFileModel.simples == null ? "" : siciFileModel.simples!;
+      siciFileModel.simples == null ? "" : siciFileModel.simples!;
       response.fields['controllerAliqSimplesPorc'] =
-          siciFileModel.simplesPorc == null ? "" : siciFileModel.simplesPorc!;
+      siciFileModel.simplesPorc == null ? "" : siciFileModel.simplesPorc!;
       response.fields['controllerICMS'] =
-          siciFileModel.icms == null ? "" : siciFileModel.icms!;
+      siciFileModel.icms == null ? "" : siciFileModel.icms!;
       response.fields['controllerICMSPorc'] =
-          siciFileModel.icmsPorc == null ? "" : siciFileModel.icmsPorc!;
+      siciFileModel.icmsPorc == null ? "" : siciFileModel.icmsPorc!;
       response.fields['controllerPIS'] =
-          siciFileModel.pis == null ? "" : siciFileModel.pis!;
+      siciFileModel.pis == null ? "" : siciFileModel.pis!;
       response.fields['controllerPISPorc'] =
-          siciFileModel.pisPorc == null ? "" : siciFileModel.pisPorc!;
+      siciFileModel.pisPorc == null ? "" : siciFileModel.pisPorc!;
       response.fields['controllerCOFINS'] =
-          siciFileModel.cofins == null ? "" : siciFileModel.cofins!;
+      siciFileModel.cofins == null ? "" : siciFileModel.cofins!;
       response.fields['controllerCOFINSPorc'] =
-          siciFileModel.cofinsPorc == null ? "" : siciFileModel.cofinsPorc!;
+      siciFileModel.cofinsPorc == null ? "" : siciFileModel.cofinsPorc!;
       response.fields['controllerReceitaLiquida'] =
-          siciFileModel.receitaLiquida == null
-              ? ""
-              : siciFileModel.receitaLiquida!;
+      siciFileModel.receitaLiquida == null
+          ? ""
+          : siciFileModel.receitaLiquida!;
       response.fields['controllerObservacoes'] =
           siciFileModel.observacoes ?? '';
       int index = 1;
@@ -358,7 +326,7 @@ class ServicoMobileService {
           throw (ApiRestInformation.problemOfComunication);
         } else {
           Map<String, dynamic> map =
-              jsonDecode(Components.removeAllHtmlTags(respStr));
+          jsonDecode(Components.removeAllHtmlTags(respStr));
           OperationJson resp = OperationJson.fromJson(map);
           operacao.erro = !resp.status!;
           operacao.message = resp.message;
@@ -393,15 +361,14 @@ class ServicoMobileService {
   static Future<Operation> onRecoversSiciReleases() async {
     Operation operacao = Operation();
     try {
+
       String? token = await ComponentsJWTToken.JWTTokenPadrao();
       final response = await http
-          .post(
-              Uri.parse(
-                  (kDebugMode ? Url : UrlDebug) + "/analise/recuperar_ws"),
-              headers: ApiRestInformation.onHeadersToken(token!),
-              body: null,
-              encoding: Encoding.getByName('utf-8'))
-          .timeout(const Duration(seconds: 10));
+          .post(Uri.parse("$Url/analise/recuperar_ws"),
+          headers: ApiRestInformation.onHeadersToken(token!),
+          body: null,
+          encoding: Encoding.getByName('utf-8'));
+          //.timeout(const Duration(seconds: 30));
       operacao.erro = false;
       operacao.message = 'Operação realizada com sucesso';
       operacao.result = null;
@@ -410,11 +377,10 @@ class ServicoMobileService {
         if (!response.body.isNotEmpty) {
           throw (ApiRestInformation.problemOfComunication);
         } else {
-          print(response.body);
           Map<String, dynamic> map = jsonDecode(response.body);
           OperationJson resp = OperationJson.fromJson(map);
           operacao.erro = !resp.status!;
-          operacao.message = resp.message;
+          operacao.message = resp.message ?? 'Transação realizada com sucesso';
           operacao.result = resp.result;
           operacao.resultList = resp.result as List;
         }
@@ -444,27 +410,13 @@ class ServicoMobileService {
     return operacao;
   }
 
-  static Future<Operation> onRegisterUser(
-      String nome,
-      String cpf,
-      String email,
-      String telefone,
-      String telefoneWhatsapp,
-      String empresa,
-      String uf) async {
+  static Future<Operation> onRegisterUser(String nome, String cpf, String email, String telefone, String telefoneWhatsapp, String empresa, String uf) async {
     Operation operacao = Operation();
     try {
       String? token = await ComponentsJWTToken.JWTTokenPadrao();
-      Map<String, String> headers = {
-        'Content-Type': 'application/json; charset=utf-8',
-        "token": token!,
-      };
       http.MultipartRequest response;
-      response = http.MultipartRequest(
-          'POST',
-          Uri.parse(
-              (kDebugMode ? Url : UrlDebug) + "/usuario/inserir_usuario_ws"));
-      response.headers.addAll(headers);
+      response = http.MultipartRequest('POST', Uri.parse("$Url/usuario/inserir_usuario_ws"));
+      response.headers.addAll(ApiRestInformation.onHeadersToken(token!));
       response.fields['controllerNome'] = nome;
       response.fields['controllerCPF'] = cpf;
       response.fields['controllerEmail'] = email;
@@ -475,11 +427,10 @@ class ServicoMobileService {
       var streamedResponse = await response.send();
       final respStr = await streamedResponse.stream.bytesToString();
       if (streamedResponse.statusCode == 200) {
-        if (streamedResponse.stream.isEmpty == true) {
+        if (respStr.isEmpty) {
           throw (ApiRestInformation.problemOfComunication);
         } else {
-          Map<String, dynamic> map =
-              jsonDecode(Components.removeAllHtmlTags(respStr));
+          Map<String, dynamic> map = jsonDecode(Components.removeAllHtmlTags(respStr));
           OperationJson resp = OperationJson.fromJson(map);
           operacao.erro = !resp.status!;
           operacao.message = resp.message;
@@ -489,8 +440,59 @@ class ServicoMobileService {
           throw ('Não foi identificado resposta');
         }
       } else {
-        operacao = await ApiRestStatusAnswerHTTP.AnswersHTTP(
-            streamedResponse.statusCode, streamedResponse.stream.toString());
+        operacao = await ApiRestStatusAnswerHTTP.AnswersHTTP(streamedResponse.statusCode, streamedResponse.stream.toString());
+      }
+    } on TimeoutException {
+      operacao.erro = true;
+      operacao.message = ApiRestInformation.timeoutHttp;
+    } on SocketException {
+      operacao.erro = true;
+      operacao.message = ApiRestInformation.internetSocketException;
+    } on HttpException {
+      operacao.erro = true;
+      operacao.message = ApiRestInformation.httpException;
+    } on FormatException {
+      operacao.erro = true;
+      operacao.message = ApiRestInformation.formatException;
+    } catch (e) {
+      operacao.erro = true;
+      operacao.message = e.toString();
+    }
+    return operacao;
+  }
+
+  static Future<Operation> onSiciArquivoUpload(String nmeArquivo, Uint8List imageFile) async {
+    Operation operacao = Operation();
+    try {
+      String? token = await ComponentsJWTToken.JWTTokenPadrao();
+      http.MultipartRequest response;
+      response = http.MultipartRequest('POST', Uri.parse("$Url/analise/adicionarArquivo_ws"));
+      response.headers.addAll(ApiRestInformation.onHeadersToken(token!));
+      response.fields['ambiente'] = 'dici';
+      response.files.add(
+        http.MultipartFile.fromBytes(
+          'arquivo',
+          imageFile,
+          filename: nmeArquivo,
+          contentType: MediaType("application", "xlsx"),
+        ),);
+      var streamedResponse = await response.send();
+      final respStr = await streamedResponse.stream.bytesToString();
+      if (streamedResponse.statusCode == 200) {
+        if (respStr.isEmpty) {
+          throw (ApiRestInformation.problemOfComunication);
+        } else {
+          Map<String, dynamic> map = jsonDecode(Components.removeAllHtmlTags(respStr));
+          OperationJson resp = OperationJson.fromJson(map);
+          operacao.erro = !resp.status!;
+          operacao.message = resp.message ?? 'Transação realizada com sucesso';
+          operacao.result = map;
+        }
+        if (operacao.message == null) {
+          throw ('Não foi identificado resposta');
+        }
+      } else {
+        operacao = await ApiRestStatusAnswerHTTP.AnswersHTTP(streamedResponse.statusCode, streamedResponse.stream.toString());
       }
     } on TimeoutException {
       operacao.erro = true;
@@ -515,24 +517,16 @@ class ServicoMobileService {
     Operation operacao = Operation();
     try {
       String? token = await ComponentsJWTToken.JWTTokenPadrao();
-      Map<String, String> headers = {
-        'Content-Type': 'application/json; charset=utf-8',
-        'token': token!,
-      };
       http.MultipartRequest response;
-      response = http.MultipartRequest(
-          'POST',
-          Uri.parse((kDebugMode ? Url : UrlDebug) +
-              "/notificacoes/Notificacoes_ws/recuperarNotificacao_ws"));
-      response.headers.addAll(headers);
+      response = http.MultipartRequest('POST', Uri.parse("$Url/notificacoes/Notificacoes_ws/recuperarNotificacao_ws"));
+      response.headers.addAll(ApiRestInformation.onHeadersToken(token!));
       var streamedResponse = await response.send();
       final respStr = await streamedResponse.stream.bytesToString();
       if (streamedResponse.statusCode == 200) {
-        if (streamedResponse.stream.isEmpty == true) {
+        if (respStr.isEmpty) {
           throw (ApiRestInformation.problemOfComunication);
         } else {
-          Map<String, dynamic> map =
-              jsonDecode(Components.removeAllHtmlTags(respStr));
+          Map<String, dynamic> map = jsonDecode(Components.removeAllHtmlTags(respStr));
           OperationJson resp = OperationJson.fromJson(map);
           operacao.erro = !resp.status!;
           operacao.message = resp.message;
@@ -542,8 +536,7 @@ class ServicoMobileService {
           throw ('Não foi identificado resposta');
         }
       } else {
-        operacao = await ApiRestStatusAnswerHTTP.AnswersHTTP(
-            streamedResponse.statusCode, streamedResponse.stream.toString());
+        operacao = await ApiRestStatusAnswerHTTP.AnswersHTTP(streamedResponse.statusCode, streamedResponse.stream.toString());
       }
     } on TimeoutException {
       operacao.erro = true;
@@ -568,25 +561,20 @@ class ServicoMobileService {
     Operation operacao = Operation();
     try {
       String? token = await ComponentsJWTToken.JWTTokenPadrao();
-      Map<String, String> headers = {
-        'Content-Type': 'application/json; charset=utf-8',
-        'token': token!,
-      };
       http.MultipartRequest response;
       response = http.MultipartRequest(
           'POST',
-          Uri.parse((kDebugMode ? Url : UrlDebug) +
-              "/notificacoes/Notificacoes_ws/recuperarNotificacao_ws"));
-      response.headers.addAll(headers);
+          Uri.parse(
+              "$Url/notificacoes/Notificacoes_ws/recuperarNotificacao_ws"));
+      response.headers.addAll(ApiRestInformation.onHeadersToken(token!));
       response.fields['id'] = idUser;
       var streamedResponse = await response.send();
       final respStr = await streamedResponse.stream.bytesToString();
       if (streamedResponse.statusCode == 200) {
-        if (streamedResponse.stream.isEmpty == true) {
+        if (respStr.isEmpty) {
           throw (ApiRestInformation.problemOfComunication);
         } else {
-          Map<String, dynamic> map =
-              jsonDecode(Components.removeAllHtmlTags(respStr));
+          Map<String, dynamic> map = jsonDecode(Components.removeAllHtmlTags(respStr));
           OperationJson resp = OperationJson.fromJson(map);
           operacao.erro = !resp.status!;
           operacao.message = resp.message;
@@ -618,30 +606,25 @@ class ServicoMobileService {
     return operacao;
   }
 
-  static Future<Operation> onRecuperaNotificacaoPeloId(
-      String IdNotificacao) async {
+  static Future<Operation> onRecuperaNotificacaoPeloId(String IdNotificacao) async {
     Operation operacao = Operation();
     try {
       String? token = await ComponentsJWTToken.JWTTokenPadrao();
-      Map<String, String> headers = {
-        'Content-Type': 'application/json; charset=utf-8',
-        'token': token!,
-      };
       http.MultipartRequest response;
       response = http.MultipartRequest(
           'POST',
-          Uri.parse((kDebugMode ? Url : UrlDebug) +
-              "/notificacoes/Notificacoes_ws/recuperarNotificacao_ws"));
-      response.headers.addAll(headers);
+          Uri.parse(
+              "$Url/notificacoes/Notificacoes_ws/recuperarNotificacao_ws"));
+      response.headers.addAll(ApiRestInformation.onHeadersToken(token!));
       response.fields['id'] = IdNotificacao;
       var streamedResponse = await response.send();
       final respStr = await streamedResponse.stream.bytesToString();
       if (streamedResponse.statusCode == 200) {
-        if (streamedResponse.stream.isEmpty == true) {
+        if (respStr.isEmpty) {
           throw (ApiRestInformation.problemOfComunication);
         } else {
           Map<String, dynamic> map =
-              jsonDecode(Components.removeAllHtmlTags(respStr));
+          jsonDecode(Components.removeAllHtmlTags(respStr));
           OperationJson resp = OperationJson.fromJson(map);
           operacao.erro = !resp.status!;
           operacao.message = resp.message;
@@ -677,25 +660,21 @@ class ServicoMobileService {
     Operation operacao = Operation();
     try {
       String? token = await ComponentsJWTToken.JWTTokenPadrao();
-      Map<String, String> headers = {
-        'Content-Type': 'application/json; charset=utf-8',
-        "token": token!,
-      };
       http.MultipartRequest response;
       response = http.MultipartRequest(
           'POST',
-          Uri.parse((kDebugMode ? Url : UrlDebug) +
-              "/notificacoes/Notificacoes_ws/recuperarNotificacao_ws"));
-      response.headers.addAll(headers);
+          Uri.parse(
+              "$Url/notificacoes/Notificacoes_ws/recuperarNotificacao_ws"));
+      response.headers.addAll(ApiRestInformation.onHeadersToken(token!));
       response.fields['id'] = idNotificacao;
       var streamedResponse = await response.send();
       final respStr = await streamedResponse.stream.bytesToString();
       if (streamedResponse.statusCode == 200) {
-        if (streamedResponse.stream.isEmpty == true) {
+        if (respStr.isEmpty) {
           throw (ApiRestInformation.problemOfComunication);
         } else {
           Map<String, dynamic> map =
-              jsonDecode(Components.removeAllHtmlTags(respStr));
+          jsonDecode(Components.removeAllHtmlTags(respStr));
           OperationJson resp = OperationJson.fromJson(map);
           operacao.erro = !resp.status!;
           operacao.message = resp.message;
@@ -731,16 +710,11 @@ class ServicoMobileService {
     Operation operacao = Operation();
     try {
       String? token = await ComponentsJWTToken.JWTTokenPadrao();
-      Map<String, String> headers = {
-        'Content-Type': 'application/json; charset=utf-8',
-        "token": token!,
-      };
+
       http.MultipartRequest response;
-      response = http.MultipartRequest(
-          'POST',
-          Uri.parse((kDebugMode ? Url : UrlDebug) +
-              "/notificacoes/recuperarTodasNotificacaoByCpf_ws"));
-      response.headers.addAll(headers);
+      response = http.MultipartRequest('POST',
+          Uri.parse("$Url/notificacoes/recuperarTodasNotificacaoByCpf_ws"));
+      response.headers.addAll(ApiRestInformation.onHeadersToken(token!));
       response.fields['cpf'] = cpf;
       var streamedResponse = await response.send();
       final respStr = await streamedResponse.stream.bytesToString();
@@ -787,16 +761,9 @@ class ServicoMobileService {
     Operation operacao = Operation();
     try {
       String? token = await ComponentsJWTToken.JWTTokenPadrao();
-      Map<String, String> headers = {
-        'Content-Type': 'application/json; charset=utf-8',
-        "token": token!,
-      };
       http.MultipartRequest response;
-      response = http.MultipartRequest(
-          'POST',
-          Uri.parse((kDebugMode ? Url : UrlDebug) +
-              "/lancamento/recuperar_lista_documentos_ws"));
-      response.headers.addAll(headers);
+      response = http.MultipartRequest('POST', Uri.parse("$Url/lancamento/recuperar_lista_documentos_ws"));
+      response.headers.addAll(ApiRestInformation.onHeadersToken(token!));
       var streamedResponse = await response.send();
       final respStr = await streamedResponse.stream.bytesToString();
       operacao.statusCode = streamedResponse.statusCode;
@@ -804,8 +771,8 @@ class ServicoMobileService {
         if (respStr.isEmpty) {
           throw (ApiRestInformation.problemOfComunication);
         } else {
-          Map<String, dynamic> map = jsonDecode(Components.removeAllHtmlTags(
-              Components.removeAllHtmlTags(respStr)));
+          Map<String, dynamic> map = jsonDecode(Components.removeAllHtmlTags(Components.removeAllHtmlTags(respStr)));
+          print(map);
           OperationJson resp = OperationJson.fromJson(map);
           operacao.erro = !resp.status!;
           operacao.message = resp.message;
@@ -842,17 +809,12 @@ class ServicoMobileService {
     Operation operacao = Operation();
     try {
       String? token = await ComponentsJWTToken.JWTTokenPadrao();
-      Map<String, String> headers = {
-        'Content-Type': 'application/json; charset=utf-8',
-        "token": token!,
-      };
       http.MultipartRequest response;
       response = http.MultipartRequest(
-          'POST',
-          Uri.parse(
-              (kDebugMode ? Url : UrlDebug) + "/recibos/download_arquivo_ws"));
-      response.headers.addAll(headers);
+          'POST', Uri.parse("$Url/recibos/download_arquivo_ws"));
+      response.headers.addAll(ApiRestInformation.onHeadersToken(token!));
       response.fields['id'] = idDocumento;
+      response.fields['base64'] = "true";
       var streamedResponse = await response.send();
       var responseData = await streamedResponse.stream.toBytes();
       var responseToString = String.fromCharCodes(responseData);
@@ -869,8 +831,7 @@ class ServicoMobileService {
           throw ('Não foi identificado resposta');
         }
       } else {
-        operacao = await ApiRestStatusAnswerHTTP.AnswersHTTP(
-            streamedResponse.statusCode, streamedResponse.stream.toString());
+        operacao = await ApiRestStatusAnswerHTTP.AnswersHTTP(streamedResponse.statusCode, streamedResponse.stream.toString());
       }
     } on TimeoutException {
       operacao.erro = true;
@@ -893,14 +854,11 @@ class ServicoMobileService {
 }
 
 class ApiRestInformation {
-  static String problemOfComunication =
-      'Houve um problema de comunicação com os servidores da scmengenharia';
-  static String internetSocketException =
-      'Verifique sua conexão com a internet e tente novamente';
+  static String problemOfComunication = 'Houve um problema de comunicação com os servidores da scmengenharia';
+  static String internetSocketException = 'Verifique sua conexão com a internet e tente novamente';
   static String httpException = 'Não foi possível encontrar a postagem';
   static String formatException = 'Formato de resposta ruim 👎';
-  static String timeoutHttp =
-      ' O tempo limite esgotou antes da conclusão da operação ou o servidor não está respondendo';
+  static String timeoutHttp = ' O tempo limite esgotou antes da conclusão da operação ou o servidor não está respondendo';
   static Map<String, String> onHeaders() {
     Map<String, String> headers = {
       'Content-Type': 'application/json; charset=utf-8',
@@ -913,12 +871,13 @@ class ApiRestInformation {
   }
 
   static Map<String, String> onHeadersToken(String token) {
+
     Map<String, String> headers = {
       'Content-Type': 'application/json; charset=utf-8',
       'Accept': 'application/json',
       'Access-Control-Allow-Methods': '*',
       'Access-Control-Allow-Headers': '*',
-      'token': token!,
+      'token': token,
     };
     return headers;
   }
@@ -928,12 +887,12 @@ class ApiRestStatusAnswerHTTP {
   static Future<Operation> AnswersHTTP(int statusCode, String body) async {
     Operation operation = Operation();
     switch (statusCode) {
-      //Erro do cliente
+    //Erro do cliente
       case 400:
         {
           operation.erro = true;
           operation.message =
-              'O servidor não pode ou não processará a solicitação';
+          'O servidor não pode ou não processará a solicitação';
           operation.result = null;
           operation.statusCode = 400;
         }
@@ -942,7 +901,7 @@ class ApiRestStatusAnswerHTTP {
         {
           operation.erro = true;
           operation.message =
-              'O servidor não pode ou não processará a solicitação';
+          'O servidor não pode ou não processará a solicitação';
           operation.result = null;
           operation.statusCode = 409;
           try {} catch (error) {
@@ -950,12 +909,12 @@ class ApiRestStatusAnswerHTTP {
           }
         }
         break;
-      //Respostas de erro do Servidor
+    //Respostas de erro do Servidor
       case 500:
         {
           operation.erro = true;
           operation.message =
-              'O servidor encontrou uma condição inesperada que o impediu de atender à solicitação.  (ERRO DO SERVIDOR)';
+          'O servidor encontrou uma condição inesperada que o impediu de atender à solicitação.  (ERRO DO SERVIDOR)';
           operation.result = null;
           operation.statusCode = 500;
           try {
@@ -971,7 +930,7 @@ class ApiRestStatusAnswerHTTP {
         {
           operation.erro = true;
           operation.message =
-              'O servidor não oferece suporte à funcionalidade necessária para atender à solicitação. (ERRO DO SERVIDOR)';
+          'O servidor não oferece suporte à funcionalidade necessária para atender à solicitação. (ERRO DO SERVIDOR)';
           operation.result = null;
           operation.statusCode = 501;
         }
@@ -980,7 +939,7 @@ class ApiRestStatusAnswerHTTP {
         {
           operation.erro = true;
           operation.message =
-              'O servidor, ao atuar como gateway ou proxy, recebeu uma resposta inválida. (ERRO DO SERVIDOR)'; //BAD GATEWAY
+          'O servidor, ao atuar como gateway ou proxy, recebeu uma resposta inválida. (ERRO DO SERVIDOR)'; //BAD GATEWAY
           operation.result = null;
           operation.statusCode = 502;
         }
@@ -989,7 +948,7 @@ class ApiRestStatusAnswerHTTP {
         {
           operation.erro = true;
           operation.message =
-              'No momento, o servidor não pode atender à solicitação devido a uma sobrecarga temporária ou manutenção programada. (ERRO DO SERVIDOR)'; //SERVIÇO INDISPONÍVEL
+          'No momento, o servidor não pode atender à solicitação devido a uma sobrecarga temporária ou manutenção programada. (ERRO DO SERVIDOR)'; //SERVIÇO INDISPONÍVEL
           operation.result = null;
           operation.statusCode = 503;
         }
@@ -998,7 +957,7 @@ class ApiRestStatusAnswerHTTP {
         {
           operation.erro = true;
           operation.message =
-              'O servidor, ao atuar como gateway ou proxy, não recebeu uma resposta oportuna. (ERRO DO SERVIDOR)'; //GATEWAY TIMEOUT
+          'O servidor, ao atuar como gateway ou proxy, não recebeu uma resposta oportuna. (ERRO DO SERVIDOR)'; //GATEWAY TIMEOUT
           operation.result = null;
           operation.statusCode = 504;
         }
@@ -1007,7 +966,7 @@ class ApiRestStatusAnswerHTTP {
         {
           operation.erro = true;
           operation.message =
-              'O servidor não oferece suporte ou se recusa a oferecer suporte à versão principal de HTTP. (ERRO DO SERVIDOR)'; //VERSÃO HTTP NÃO SUPORTADA
+          'O servidor não oferece suporte ou se recusa a oferecer suporte à versão principal de HTTP. (ERRO DO SERVIDOR)'; //VERSÃO HTTP NÃO SUPORTADA
           operation.result = null;
           operation.statusCode = 505;
         }
@@ -1016,7 +975,7 @@ class ApiRestStatusAnswerHTTP {
         {
           operation.erro = true;
           operation.message =
-              'O servidor tem um erro de configuração interno. (ERRO DO SERVIDOR)';
+          'O servidor tem um erro de configuração interno. (ERRO DO SERVIDOR)';
           operation.result = null;
           operation.statusCode = 506;
         }
@@ -1025,7 +984,7 @@ class ApiRestStatusAnswerHTTP {
         {
           operation.erro = true;
           operation.message =
-              'O método não pôde ser executado no recurso porque o servidor não pode armazenar a representação necessária para concluir a solicitação com êxito. (ERRO DO SERVIDOR)'; //ARMAZENAMENTO INSUFICIENTE
+          'O método não pôde ser executado no recurso porque o servidor não pode armazenar a representação necessária para concluir a solicitação com êxito. (ERRO DO SERVIDOR)'; //ARMAZENAMENTO INSUFICIENTE
           operation.result = null;
           operation.statusCode = 507;
         }
@@ -1034,7 +993,7 @@ class ApiRestStatusAnswerHTTP {
         {
           operation.erro = true;
           operation.message =
-              'O servidor encerrou uma operação porque encontrou um loop infinito ao processar uma solicitação. (ERRO DO SERVIDOR)'; //LOOP DETECTADO
+          'O servidor encerrou uma operação porque encontrou um loop infinito ao processar uma solicitação. (ERRO DO SERVIDOR)'; //LOOP DETECTADO
           operation.result = null;
           operation.statusCode = 508;
         }
@@ -1043,7 +1002,7 @@ class ApiRestStatusAnswerHTTP {
         {
           operation.erro = true;
           operation.message =
-              'A política de acesso ao recurso não foi atendida na solicitação. (ERRO DO SERVIDOR)'; //NÃO ESTENDIDO
+          'A política de acesso ao recurso não foi atendida na solicitação. (ERRO DO SERVIDOR)'; //NÃO ESTENDIDO
           operation.result = null;
           operation.statusCode = 510;
         }
@@ -1052,7 +1011,7 @@ class ApiRestStatusAnswerHTTP {
         {
           operation.erro = true;
           operation.message =
-              'É preciso se autenticar para obter acesso à rede. (ERRO DO SERVIDOR)'; // AUTENTICAÇÃO DE REDE NECESSÁRIA
+          'É preciso se autenticar para obter acesso à rede. (ERRO DO SERVIDOR)'; // AUTENTICAÇÃO DE REDE NECESSÁRIA
           operation.result = null;
           operation.statusCode = 511;
         }
@@ -1061,7 +1020,7 @@ class ApiRestStatusAnswerHTTP {
         {
           operation.erro = true;
           operation.message =
-              'Erro de tempo limite de conexão de rede. (ERRO DO SERVIDOR)'; //ERRO DE TEMPO LIMITE DE CONEXÃO DA REDE
+          'Erro de tempo limite de conexão de rede. (ERRO DO SERVIDOR)'; //ERRO DE TEMPO LIMITE DE CONEXÃO DA REDE
           operation.result = null;
           operation.statusCode = 599;
         }
